+++ Common/src/main/java/net/potionstudios/biomeswevegone/world/level/levelgen/biome/BWGOverworldBiomes.java
@@ -504,7 +504,6 @@ class BWGOverworldBiomes {
         addVegetal(generationSettings, AquaticPlacements.WARM_OCEAN_VEGETATION);
         addVegetal(generationSettings, AquaticPlacements.SEAGRASS_WARM);
         addVegetal(generationSettings, AquaticPlacements.SEA_PICKLE);
-        addVegetal(generationSettings, AquaticPlacements.SEAGRASS_NORMAL);
         BiomeDefaultFeatures.addLushCavesVegetationFeatures(generationSettings);
 
         MobSpawnSettings.Builder spawnSettings = new MobSpawnSettings.Builder();