+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 1.3.3
 - Make Entity Model Classes Package-Private, remove pointless override of getAnimation Method
 - Decrease Pumpkin Warden Shadow Radius
+- Fix Incompatibility with William Wythers' Overhauled Overworld/Expanded Ecosphere
+  - Remove SeaGrass Normal from Lush Stacks as SeaGrass Warm is the same feature but more common 
 
 # 1.3.2
 - Fix and Add Missing Mossy Red Rock and Mossy Stone Recipes